Update all markdown links to instruction to a specific *.md page
As part of this exercise, we searched all files with the following
regex to ensure that we caught all instances of the problem.

(?<!!)\[.{1,40}?\]\((?!http|\w+-code).{1,50}?(?<!.md|.jpg|.java|starter-code|readme|.txt)\)

This started by searching for all markdown links, and then we excluded
multiple kinds of acceptable links until we narrowed down the search results.

After fixing the errors contained in this commit, the only remaining result
in the regex is a corner case to the location `.`. Because this, we can be sure
that we've found all of the occurences of this issue.

The only potentially questional assumption we've made here is that links to
code sometimes intend to reference the parent directory. That seems reasonable,
but if that isn't always the intention, we can obviously fix it.