Remove indications that each feature is worth individual points...
Originally, I attempted to solve this by listing all of the features
in a bulleted list in a single row of the table. This most accurately
reflects the grading realities; however, it is very difficult to maintain
since markdown files do not have builtin support for lists within tables.
The workaround requires using HTML tags (which is fine), but then they all
have to be condensed into a single line for the markdown to render properly.

In this approach, we remove the points from each item and indicate that they
are sub-topics of the parent item which is 'Program Functionality'.
I chose the abbreviation 'feat:' because it is short and conveys the intention
of the full word 'functionality.'

I considered using ASCII characters to create a directory-like structure,
but I figured that wouldn't actually work well from separate table rows.