Added Phase 5 Passoff Common Problems
Added the information for the Phase 5 Passoff into Chess Tips, and included information for 4 different common problems. Make sure register signs you in, that list games numbers the game independent of gameID, that you print the chess board correctly, and print readable errors and giving the example of joining or observing a game with an invalid gameID as one that students should make sure works smoothly.